refactor: reorganize renderer and services
