--- src/ink-renderer/index.ts
@@ -1,1 +1,0 @@
-export { InkRenderer, type InkRendererService } from "./service";
