+++ src/ink-renderer/snapshot/types.ts
@@ -1,4 +1,4 @@
-import type { TaskSnapshot } from "../types";
+import type { TaskSnapshot } from "../../types";
 
 export interface TaskTreeInfo {
   readonly depth: number;
