+++ src/ink-renderer/view/hooks/use-spinner-clock.ts
