+++ src/ink-renderer/view/tree-prefix.ts
@@ -1,0 +1,16 @@
+import type { TaskTreeInfo } from "../snapshot/types";
+
+const treeAncestorPrefix = (tree: TaskTreeInfo): string =>
+  tree.ancestorHasNextSibling
+    .slice(1)
+    .map((hasNext) => (hasNext ? "│  " : "   "))
+    .join("");
+
+export const renderTreePrefix = (tree: TaskTreeInfo): string => {
+  if (tree.depth <= 0) {
+    return "";
+  }
+
+  const ancestor = treeAncestorPrefix(tree);
+  return `${ancestor}${tree.hasNextSibling ? "├─ " : "└─ "}`;
+};
