+++ src/services/index.ts
@@ -1,0 +1,3 @@
+export { InkRenderer, type InkRendererService } from "./ink-renderer";
+export { Progress } from "./progress";
+export { ProgressStdio, type ProgressStdioService } from "./stdio";
